minor clean up of types and utils
